install Laravel Breeze and implement proper authentication

+++ app/Http/Controllers/api/v1/PostController.php
@@ -7,6 +7,8 @@ use App\Http\Requests\StorePostRequest;
 use Illuminate\Http\Request;
 use App\Models\post;
 use App\Http\Resources\PostResource;
+use Illuminate\Container\Attributes\Auth;
+use Illuminate\Support\Facades\Auth as FacadesAuth;
 
 class PostController extends Controller
 {
@@ -15,7 +17,9 @@ class PostController extends Controller
      */
     public function index()
     {
-        return PostResource::collection(Post::with('author')->paginate());
+        $user = request()->user();
+        $posts = $user->posts()->paginate();
+        return PostResource::collection($posts);
     }
 
     /**
@@ -24,7 +28,7 @@ class PostController extends Controller
     public function store(StorePostRequest $request)
     {
         $data = $request->validated();
-        $data['author_id'] = 1;
+        $data['author_id'] = $request->user()->id;
 
         $post = Post::create($data);
 
@@ -36,6 +40,8 @@ class PostController extends Controller
      */
     public function show(Post $post)
     {
+        abort_if(FacadesAuth::id() != $post->author_id, 403, 'Access Forbidden');
+
         return new PostResource($post);
     }
 
@@ -44,6 +50,7 @@ class PostController extends Controller
      */
     public function update(Request $request, Post $post)
     {
+        abort_if(FacadesAuth::id() != $post->author_id, 403, 'Access Forbidden');
         $data = $request->validate([
             'title' => 'required|string|min:2',
             'body' => ['required', 'string', 'min:2']
@@ -59,6 +66,7 @@ class PostController extends Controller
      */
     public function destroy(Post $post)
     {
+        abort_if(FacadesAuth::id() != $post->author_id, 403, 'Access Forbidden');
         $post->delete();
         return response()->noContent();
     }
